Add spin symbol via the :spin token
Supported in both the R and C++ interfaces.

This is useful where either the number of things in the progress bar
is much larger than the width of the bar, where the total number is
unknown, or (my main use case) where many calls to tick are `tick(0)`.

With :spin, the progress bar will look like:

```
(\) [===----...
(|) [====---...
(/) [=====--...
(-) [======-...
```